+++ imednet/testing/fake_data.py
@@ -6,9 +6,10 @@ from typing import Any, Dict, List, Optional
 
 from faker import Faker
 
+from imednet.validation.cache import SchemaCache
+
 from ..api.models.forms import Form
 from ..api.models.variables import Variable
-from imednet.validation.cache import SchemaCache
 
 faker = Faker()